% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditionalModel.R
\name{traditionalModel}
\alias{traditionalModel}
\title{Specify and fit model using count data from traditional (non-eDNA) surveys}
\usage{
traditionalModel(
  data,
  family = "poisson",
  q = FALSE,
  n.chain = 4,
  n.iter.burn = 500,
  n.iter.sample = 2500,
  adapt_delta = 0.9
)
}
\arguments{
\item{data}{A list containing data necessary for model fitting. Valid tags are \code{count} and \code{count.type}. \code{count} is a matrix or data frame of traditional survey count data, with first dimension equal to the number of sites (i) and second dimension equal to the maximum number of traditional survey replicates at a given site (j). \code{count.type} is an optional matrix or data frame of integers indicating gear type (k) used in corresponding count data, with first dimension equal to the number of sites (i) and second dimension equal to the maximum number of traditional survey replicates at a given site (j). Values should be integers beginning with 1 (referring to the first gear type) to n (last gear type). Empty cells should be NA. Sites, i, should be consistent in all count data.}

\item{family}{The distribution class used to model traditional survey count data. Options include poisson ('poisson') and negative binomial ('negbin'). Default value is 'poisson'.}

\item{q}{A logical value indicating whether to estimate a catchability coefficient, q, for traditional survey gear types (TRUE) or to not estimate a catchability coefficient, q, for traditional survey gear types (FALSE). Default value is FALSE.}

\item{n.chain}{Number of MCMC chains. Default value is 4.}

\item{n.iter.burn}{Number of warm-up MCMC iterations. Default value is 500.}

\item{n.iter.sample}{Number of sampling MCMC iterations. Default value is 2500.}

\item{adapt_delta}{Target average acceptance probability used in \code{rstan::sampling}. Default value is 0.9.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
This function implements a Bayesian model that estimates expected species catch rate using count data from traditional (non-eDNA surveys). When multiple traditional gear types are used, an optional variation allows estimation of catchability coefficients, which scale the catchability of gear types relative to the expected catch rate of a reference gear type. Model is implemented using Bayesian inference using the \code{rstan} package, which uses Hamilton Monte Carlo to simulate the posterior distributions.
}
\note{
Before fitting the model, this function checks to ensure that the model specification is possible given the data files. These checks include:
\itemize{
\item  All tags in data are valid (i.e., include count and count.type).
\item  Number of sites in count and count type data are equal.
\item  All data are numeric (i.e., integer or NA).
\item  Empty data cells (NA) match in count and count.type.
\item  family is either 'poisson' or 'negbin'.
\item  q_ref is an integer >= 0.
}

If any of these checks fail, the function returns an error message.
}
\examples{
\donttest{
data(greencrabData)

# Examine data in list
# This function uses only traditional survey count data and optionally the count type data
names(greencrabData)

# Note that the surveyed sites (rows) should match in the data
dim(greencrabData$count)[1]
dim(greencrabData$count.type)[1]

# Fit a model without estimating a catchability coefficient for traditional survey gear types.
# This model assumes all traditional survey methods have the same catchability.
# Count data is modeled using a poisson distribution.
fit.no.q = traditionalModel(data=greencrabData, family='poisson', q=FALSE)


# Fit a model estimating a catchability coefficient for traditional survey gear types.
# This model does not assume all traditional survey methods have the same catchability.
# Count data is modeled using a negative binomial distribution.
fit.q = traditionalModel(data=greencrabData, family='negbin', q=TRUE)
}

}
