% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNAjoint-package.R
\docType{package}
\name{eDNAjoint-package}
\alias{eDNAjoint-package}
\alias{eDNAjoint}
\title{The 'eDNAjoint' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
