% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNAjoint-package.R
\docType{_PACKAGE}
\name{eDNAjoint-package}
\alias{eDNAjoint-package}
\alias{eDNAjoint}
\title{The 'eDNAjoint' package.}
\description{
Models integrate environmental DNA (eDNA) detection data and traditional survey data to jointly estimate species catch rate. Models can be used with count data via traditional survey methods (i.e., trapping, electrofishing, visual) and replicated eDNA detection/nondetection data via polymerase chain reaction (i.e., PCR or qPCR) from multiple survey locations. Estimated parameters include probability of a false positive eDNA detection, a site-level covariates that scale the sensitivity of eDNA surveys relative to traditional surveys, and catchability coefficients for traditional gear types. Models are implemented with a Bayesian framework (Markov chain Monte Carlo) using the 'Stan' probabilistic programming language.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version NA. https://mc-stan.org
}
